Moved all smarthome functionality into smart/ subfolder
