+++ lib/smart/connection.dart
