+++ lib/smart/connection_card.dart
@@ -17,7 +17,9 @@
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import './common.dart';
+
+import '../common.dart';
+
 import './iot_web.dart';
 import './connection.dart';
