+++ lib/smart/iot_web.dart
@@ -15,15 +15,13 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-// Documentation web views
-
 import 'package:flutter/material.dart';
+
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
-// import 'package:platform_device_id/platform_device_id.dart';
 import 'package:url_launcher/url_launcher.dart' show launchUrl, LaunchMode;
 
-import './theme.dart' show standardAppBar;
-import './common.dart' show dlog;
+import '../theme.dart' show standardAppBar;
+import '../common.dart' show dlog;
 
 const String kDocsDir = 'docs';
 const String kLoadingHTMLFilePath = "$kDocsDir/loading.html";
