+++ lib/smart/mdns_scan.dart
@@ -19,11 +19,13 @@
 
 import 'package:flutter/foundation.dart' show kReleaseMode;
 import 'package:flutter/material.dart';
+
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
-import './util.dart';
-import './theme.dart';
+import '../util.dart';
+import '../theme.dart';
+
 import './connection.dart';
 import './connection_card.dart';
 import './mdns.dart';
